% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_spec.R
\name{as.data.frame.WebSpec}
\alias{as.data.frame.WebSpec}
\title{Extract data from a WebSpec}
\usage{
\method{as.data.frame}{WebSpec}(x, ...)
}
\arguments{
\item{x}{A WebSpec object}

\item{...}{Ignored}
}
\value{
A data.frame
}
\description{
Extract data from a WebSpec
}
