% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_forest.R
\name{build_tree_level}
\alias{build_tree_level}
\title{Recursive helper for building tree levels}
\usage{
build_tree_level(split_vars, split_combos, level, parent_path)
}
\description{
Recursive helper for building tree levels
}
\keyword{internal}
