% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-components.R
\name{col_group}
\alias{col_group}
\title{Create a column group}
\usage{
col_group(header, ..., position = c("left", "right"))
}
\arguments{
\item{header}{Display header for the group}

\item{...}{Column specifications (ColumnSpec objects)}

\item{position}{Column position: "left" or "right" of the forest plot}
}
\value{
A ColumnGroup object
}
\description{
Groups multiple columns under a shared header for hierarchical display.
}
