% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-components.R
\name{col_text}
\alias{col_text}
\title{Column helper: Text column}
\usage{
col_text(field, header = NULL, ...)
}
\arguments{
\item{field}{Field name}

\item{header}{Column header}

\item{...}{Additional arguments passed to web_col}
}
\value{
A ColumnSpec object
}
\description{
Column helper: Text column
}
