% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-components.R
\name{ColumnGroup}
\alias{ColumnGroup}
\title{ColumnGroup: A group of columns with a shared header}
\usage{
ColumnGroup(
  id = character(0),
  header = character(0),
  columns = list(),
  position = "left"
)
}
\arguments{
\item{id}{Unique identifier for the group}

\item{header}{Display header text for the group}

\item{columns}{List of ColumnSpec objects in this group}

\item{position}{Column position: "left" or "right" of the forest plot}
}
\description{
Used for creating hierarchical column headers.
}
