% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-serialize.R
\name{compute_group_depths}
\alias{compute_group_depths}
\title{Compute depths for all groups based on parent hierarchy}
\usage{
compute_group_depths(groups)
}
\description{
Compute depths for all groups based on parent hierarchy
}
\keyword{internal}
