% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-annotations.R
\name{forest_annotation}
\alias{forest_annotation}
\title{Create a custom annotation}
\usage{
forest_annotation(
  study_id,
  shape = c("circle", "square", "triangle", "star"),
  position = c("after", "before", "overlay"),
  color = "#8b5cf6",
  size = 1
)
}
\arguments{
\item{study_id}{Study to annotate}

\item{shape}{Shape type}

\item{position}{Position}

\item{color}{Color}

\item{size}{Size multiplier}
}
\value{
A CustomAnnotation object
}
\description{
Create a custom annotation
}
