% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{forest_clear_filter}
\alias{forest_clear_filter}
\title{Clear filters from the forest plot}
\usage{
forest_clear_filter(proxy)
}
\arguments{
\item{proxy}{A forest_proxy object}
}
\value{
The proxy object (invisibly), for chaining
}
\description{
Clear filters from the forest plot
}
