% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{forest_filter}
\alias{forest_filter}
\title{Apply a filter to the forest plot}
\usage{
forest_filter(
  proxy,
  field,
  operator = c("eq", "neq", "gt", "lt", "contains"),
  value
)
}
\arguments{
\item{proxy}{A forest_proxy object}

\item{field}{Field to filter on}

\item{operator}{Filter operator: "eq", "neq", "gt", "lt", "contains"}

\item{value}{Filter value}
}
\value{
The proxy object (invisibly), for chaining
}
\description{
Apply a filter to the forest plot
}
