% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-annotations.R
\name{forest_rob}
\alias{forest_rob}
\title{Create a risk of bias annotation}
\usage{
forest_rob(data, study_id = "study_id", ...)
}
\arguments{
\item{data}{Data frame with study_id and domain columns}

\item{study_id}{Column name for study IDs}

\item{...}{Domain columns (unquoted)}
}
\value{
A RiskOfBias object
}
\description{
Create a risk of bias annotation
}
