% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{forest_sort}
\alias{forest_sort}
\title{Sort the forest plot by a column}
\usage{
forest_sort(proxy, column, direction = c("asc", "desc", "none"))
}
\arguments{
\item{proxy}{A forest_proxy object}

\item{column}{Column to sort by}

\item{direction}{Sort direction: "asc", "desc", or "none"}
}
\value{
The proxy object (invisibly), for chaining
}
\description{
Sort the forest plot by a column
}
