% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{forest_update_data}
\alias{forest_update_data}
\title{Update forest plot data via proxy}
\usage{
forest_update_data(proxy, data, ...)
}
\arguments{
\item{proxy}{A forest_proxy object}

\item{data}{New data (same format as forest_plot)}

\item{...}{Additional arguments passed to coerce_to_studies}
}
\value{
The proxy object (invisibly), for chaining
}
\description{
Update forest plot data via proxy
}
