% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{forestProxy}
\alias{forestProxy}
\title{Create a forest plot proxy object}
\usage{
forestProxy(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{The widget ID}

\item{session}{The Shiny session (default: current session)}
}
\value{
A forest_proxy object
}
\description{
Creates a proxy object that can be used to update a forest plot in a Shiny
app without re-rendering the entire widget.
}
