% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-core.R
\name{Heterogeneity}
\alias{Heterogeneity}
\title{Heterogeneity: Meta-analysis heterogeneity statistics}
\usage{
Heterogeneity(
  i2 = NA_real_,
  q = NA_real_,
  q_pvalue = NA_real_,
  tau2 = NA_real_
)
}
\arguments{
\item{i2}{I-squared statistic (0-100)}

\item{q}{Cochran's Q statistic}

\item{q_pvalue}{P-value for Q statistic}

\item{tau2}{Tau-squared (between-study variance)}
}
\description{
Optional metadata for meta-analysis use cases.
}
