% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_forest.R
\name{nice_linear_domain}
\alias{nice_linear_domain}
\title{Compute nice bounds for linear scale using Extended Wilkinson approach}
\usage{
nice_linear_domain(domain)
}
\description{
Compute nice bounds for linear scale using Extended Wilkinson approach
}
\keyword{internal}
