% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_forest.R
\docType{data}
\name{NICE_LOG_VALUES}
\alias{NICE_LOG_VALUES}
\title{Nice values for log scale (matches JS NICE_LOG_VALUES)}
\format{
An object of class \code{numeric} of length 48.
}
\usage{
NICE_LOG_VALUES
}
\description{
Nice values for log scale (matches JS NICE_LOG_VALUES)
}
\keyword{internal}
