% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_forest.R
\docType{data}
\name{NICE_Q}
\alias{NICE_Q}
\title{Extended Wilkinson Q sequence for "nice" tick values}
\format{
An object of class \code{numeric} of length 6.
}
\usage{
NICE_Q
}
\description{
Extended Wilkinson Q sequence for "nice" tick values
}
\keyword{internal}
