% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderForest}
\alias{renderForest}
\title{Shiny render function for forest plot}
\usage{
renderForest(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that returns a forest plot (from \code{forest_plot()})}

\item{env}{The environment in which to evaluate expr}

\item{quoted}{Is expr a quoted expression?}
}
\value{
A Shiny render function
}
\description{
Shiny render function for forest plot
}
