% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderSplitForest}
\alias{renderSplitForest}
\title{Shiny render function for split forest plot}
\usage{
renderSplitForest(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that returns a split forest plot}

\item{env}{The environment in which to evaluate expr}

\item{quoted}{Is expr a quoted expression?}
}
\value{
A Shiny render function
}
\description{
Renders a split forest plot in a Shiny application. The expression should
return either a SplitForest object or a forest_plot() call with split_by.
}
\examples{
\dontrun{
# In server function:
output$split_plot <- renderSplitForest({
  data |>
    tabviz(label = "study", columns = list(
      viz_forest(point = "or", lower = "lower", upper = "upper")
    )) |>
    split_table(by = input$split_var) |>
    forest_plot()
})
}

}
