% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-annotations.R
\name{RiskOfBias}
\alias{RiskOfBias}
\title{RiskOfBias: Risk of bias traffic light annotation}
\usage{
RiskOfBias(domains = character(0), assessments = list())
}
\arguments{
\item{domains}{Character vector of domain names}

\item{assessments}{List of RiskOfBiasAssessment objects}
}
\description{
RiskOfBias: Risk of bias traffic light annotation
}
