% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-serialize.R
\name{serialize_nav_tree}
\alias{serialize_nav_tree}
\title{Serialize navigation tree for JSON}
\usage{
serialize_nav_tree(tree)
}
\arguments{
\item{tree}{The navigation tree (list of nodes)}
}
\value{
List structure ready for JSON serialization
}
\description{
Recursively converts the R list structure to JSON-compatible format.
}
\keyword{internal}
