% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-serialize.R
\name{serialize_spec}
\alias{serialize_spec}
\title{Serialize a WebSpec to a JSON-ready list}
\usage{
serialize_spec(spec, include_forest = TRUE)
}
\arguments{
\item{spec}{A WebSpec object}

\item{include_forest}{Whether to include forest plot data (TRUE for forest_plot, FALSE for webtable)}
}
\value{
A nested list suitable for jsonlite::toJSON
}
\description{
Serialize a WebSpec to a JSON-ready list
}
\keyword{internal}
