% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-theme.R
\name{set_layout}
\alias{set_layout}
\title{Modify theme layout configuration}
\usage{
set_layout(theme, ...)
}
\arguments{
\item{theme}{A WebTheme object}

\item{...}{Named layout config values to override}
}
\value{
Modified WebTheme object
}
\description{
Pipe-friendly function to modify layout settings.
}
\examples{
web_theme_default() |>
  set_layout(plot_position = "left", cell_padding_x = 12)
}
