% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-theme.R
\name{set_shapes}
\alias{set_shapes}
\title{Modify theme shapes}
\usage{
set_shapes(theme, ...)
}
\arguments{
\item{theme}{A WebTheme object}

\item{...}{Named shape values to override}
}
\value{
Modified WebTheme object
}
\description{
Pipe-friendly function to modify shape settings.
}
\examples{
web_theme_default() |>
  set_shapes(point_size = 8, line_width = 2)
}
