% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{set_theme}
\alias{set_theme}
\title{Set theme on a WebSpec}
\usage{
set_theme(x, theme)
}
\arguments{
\item{x}{A WebSpec object or an htmlwidget created by forest_plot/webtable}

\item{theme}{Either a WebTheme object or a string matching a built-in theme name:
"default", "minimal", "dark", "jama", "lancet", "modern", "presentation",
"cochrane", or "nature"}
}
\value{
The modified WebSpec object (or widget)
}
\description{
Provides a fluent API for setting or changing the theme on a WebSpec or widget.
Accepts either a theme name (string) or a WebTheme object.
}
\examples{
\dontrun{
# Using a theme name
forest_data |>
  web_spec(hr, lower, upper) |>
  set_theme("jama") |>
  forest_plot()

# Using a WebTheme object
my_theme <- web_theme_jama() |>
  set_colors(interval = "#0066cc")

forest_data |>
  web_spec(hr, lower, upper) |>
  set_theme(my_theme) |>
  forest_plot()
}

}
