% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-theme.R
\name{set_typography}
\alias{set_typography}
\title{Modify theme typography}
\usage{
set_typography(theme, ...)
}
\arguments{
\item{theme}{A WebTheme object}

\item{...}{Named typography values to override}
}
\value{
Modified WebTheme object
}
\description{
Pipe-friendly function to modify typography settings.
}
\examples{
web_theme_default() |>
  set_typography(font_family = "Arial", font_size_base = "12pt")
}
