% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{splitForestOutput}
\alias{splitForestOutput}
\title{Shiny output function for split forest plot}
\usage{
splitForestOutput(outputId, width = "100\%", height = "600px")
}
\arguments{
\item{outputId}{Output variable name}

\item{width}{Widget width (CSS units)}

\item{height}{Widget height (CSS units)}
}
\value{
A Shiny output element
}
\description{
Creates a Shiny output element for displaying a split forest plot with
sidebar navigation.
}
