% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{splitForestProxy}
\alias{splitForestProxy}
\title{Create a split forest plot proxy object}
\usage{
splitForestProxy(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{The widget ID}

\item{session}{The Shiny session (default: current session)}
}
\value{
A split_forest_proxy object
}
\description{
Creates a proxy object that can be used to control a split forest plot in a
Shiny app without re-rendering the entire widget.
}
