% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-theme.R
\name{Typography}
\alias{Typography}
\title{Typography: Font settings}
\usage{
Typography(
  font_family = "system-ui, -apple-system, sans-serif",
  font_size_sm = "0.75rem",
  font_size_base = "0.875rem",
  font_size_lg = "1rem",
  font_weight_normal = 400,
  font_weight_medium = 500,
  font_weight_bold = 600,
  line_height = 1.5
)
}
\description{
Typography: Font settings
}
