% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_spec.R
\name{validate_forest_columns}
\alias{validate_forest_columns}
\title{Validate forest column data columns exist}
\usage{
validate_forest_columns(columns, data)
}
\arguments{
\item{columns}{List of ColumnSpec/ColumnGroup objects}

\item{data}{The data frame}
}
\description{
Checks all viz_forest() columns in the columns list (including nested groups)
and verifies their point/lower/upper columns exist in the data.
For multi-effect forest columns, validates all effect column references.
}
\keyword{internal}
