% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-components.R
\name{web_interaction}
\alias{web_interaction}
\alias{forest_interaction}
\title{Create interaction specification}
\usage{
web_interaction(
  show_filters = FALSE,
  show_legend = TRUE,
  enable_sort = TRUE,
  enable_collapse = TRUE,
  enable_select = TRUE,
  enable_hover = TRUE
)

forest_interaction(...)
}
\arguments{
\item{show_filters}{Show filter panel}

\item{show_legend}{Show legend}

\item{enable_sort}{Enable column sorting}

\item{enable_collapse}{Enable group collapsing}

\item{enable_select}{Enable row selection}

\item{enable_hover}{Enable hover effects}
}
\value{
An InteractionSpec object
}
\description{
Create interaction specification
}
