% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-theme.R
\name{web_theme_dark}
\alias{web_theme_dark}
\alias{forest_theme_dark}
\title{Create a dark theme}
\usage{
web_theme_dark()

forest_theme_dark()
}
\value{
A WebTheme object with dark mode colors
}
\description{
Create a dark theme
}
