% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-theme.R
\name{web_theme_default}
\alias{web_theme_default}
\alias{forest_theme_default}
\title{Create a default theme}
\usage{
web_theme_default()

forest_theme_default()
}
\value{
A WebTheme object with default settings
}
\description{
Create a default theme
}
