% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-theme.R
\name{web_theme_jama}
\alias{web_theme_jama}
\title{JAMA-style theme for medical journal publications}
\usage{
web_theme_jama()
}
\value{
A WebTheme object
}
\description{
Black & white, dense layout, optimized for print.
Follows JAMA (Journal of the American Medical Association) style guidelines.
}
