% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-theme.R
\name{WebTheme}
\alias{WebTheme}
\title{WebTheme: Complete theme specification}
\usage{
WebTheme(
  name = "default",
  colors = <object>,
  typography = <object>,
  spacing = <object>,
  shapes = <object>,
  axis = <object>,
  layout = <object>
)
}
\arguments{
\item{name}{Theme name}

\item{colors}{ColorPalette object}

\item{typography}{Typography object}

\item{spacing}{Spacing object}

\item{shapes}{Shapes object}

\item{axis}{AxisConfig object}

\item{layout}{LayoutConfig object}
}
\description{
WebTheme: Complete theme specification
}
