% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-components.R
\name{col_n}
\alias{col_n}
\title{Column helper: Sample size / count}
\usage{
col_n(field = "n", header = "N", width = 80, ...)
}
\arguments{
\item{field}{Field name (default "n")}

\item{header}{Column header (default "N")}

\item{width}{Column width in pixels (default 80)}

\item{...}{Additional arguments passed to web_col}
}
\value{
A ColumnSpec object
}
\description{
Column helper: Sample size / count
}
