% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-components.R
\name{col_numeric}
\alias{col_numeric}
\title{Column helper: Numeric column}
\usage{
col_numeric(field, header = NULL, ...)
}
\arguments{
\item{field}{Field name}

\item{header}{Column header}

\item{...}{Additional arguments passed to web_col}
}
\value{
A ColumnSpec object
}
\description{
Column helper: Numeric column
}
