% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-annotations.R
\name{forest_refline}
\alias{forest_refline}
\title{Create a reference line annotation}
\usage{
forest_refline(
  x,
  label = NULL,
  style = c("dashed", "solid", "dotted"),
  color = NULL
)
}
\arguments{
\item{x}{X-axis position}

\item{label}{Optional label}

\item{style}{Line style}

\item{color}{Line color}
}
\value{
A ReferenceLine object
}
\description{
Create a reference line annotation
}
