% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{forest_update_data}
\alias{forest_update_data}
\title{Update forest plot data via proxy}
\usage{
forest_update_data(proxy, spec)
}
\arguments{
\item{proxy}{A forest_proxy object}

\item{spec}{A WebSpec object (from \code{web_spec()})}
}
\value{
The proxy object (invisibly), for chaining
}
\description{
Update forest plot data via proxy
}
