% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-theme.R
\name{set_axis}
\alias{set_axis}
\title{Modify theme axis configuration}
\usage{
set_axis(theme, ...)
}
\arguments{
\item{theme}{A WebTheme object}

\item{...}{Named axis config values to override}
}
\value{
Modified WebTheme object
}
\description{
Pipe-friendly function to modify axis settings.
}
\examples{
web_theme_default() |>
  set_axis(gridlines = TRUE, range_min = 0.5, range_max = 2.0)
}
