% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-theme.R
\name{set_spacing}
\alias{set_spacing}
\title{Modify theme spacing}
\usage{
set_spacing(theme, ...)
}
\arguments{
\item{theme}{A WebTheme object}

\item{...}{Named spacing values to override}
}
\value{
Modified WebTheme object
}
\description{
Pipe-friendly function to modify spacing values.
}
\examples{
web_theme_default() |>
  set_spacing(row_height = 32, header_height = 40)
}
