% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-theme.R
\name{Shapes}
\alias{Shapes}
\title{Shapes: Shape rendering settings}
\usage{
Shapes(
  point_size = 6,
  summary_height = 10,
  line_width = 1.5,
  border_radius = 4
)
}
\description{
Shapes: Shape rendering settings
}
