% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-components.R
\name{web_col}
\alias{web_col}
\alias{forest_col}
\title{Create a column specification}
\usage{
web_col(
  field,
  header = NULL,
  type = c("text", "numeric", "interval", "bar", "pvalue", "sparkline", "custom"),
  width = NULL,
  align = NULL,
  position = c("left", "right"),
  sortable = TRUE,
  options = list()
)

forest_col(...)
}
\arguments{
\item{field}{Data field name to display}

\item{header}{Display header (defaults to field name)}

\item{type}{Column type}

\item{width}{Column width in pixels}

\item{align}{Text alignment}

\item{position}{Column position: "left" or "right" of the forest plot}

\item{sortable}{Whether sortable}

\item{options}{Named list of type-specific options}
}
\value{
A ColumnSpec object
}
\description{
Create a column specification
}
