% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-theme.R
\name{web_theme_minimal}
\alias{web_theme_minimal}
\alias{forest_theme_publication}
\title{Create a minimal/clean theme}
\usage{
web_theme_minimal()

forest_theme_publication()
}
\value{
A WebTheme object
}
\description{
Optimized for publication and print
}
