% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webforest-package.R
\docType{package}
\name{webforest-package}
\alias{webforest}
\alias{webforest-package}
\title{webforest: Web-Native Interactive Forest Plots}
\description{
Create interactive, web-native forest plots and tables. Forest plots visualize point estimates with intervals for any domain: meta-analysis, regression results, QC measurements, and more. Rendered using modern web technologies (Svelte, D3.js) for R Markdown, Quarto, Shiny, and standalone HTML. Features include collapsible groups, customizable columns, sorting, filtering, and rich theming.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kaskarn/webforest}
  \item \url{https://kaskarn.github.io/webforest/}
  \item Report bugs at \url{https://github.com/kaskarn/webforest/issues}
}

}
\author{
\strong{Maintainer}: Antoine Baldassari \email{baldassarian@gmail.com}

}
\keyword{internal}
