% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-core.R
\name{WebSpec}
\alias{WebSpec}
\title{WebSpec: Core specification for web-native table visualizations}
\arguments{
\item{data}{Processed data as a data.frame}

\item{label_col}{Column name for row labels}

\item{group_col}{Column name for grouping (optional, deepest level)}

\item{group_cols}{All group column names for hierarchical grouping (for composite ID building)}

\item{columns}{List of ColumnSpec objects defining table columns (including viz_forest)}

\item{groups}{List of GroupSpec objects}

\item{summaries}{List of GroupSummary objects}

\item{overall_summary}{Optional overall summary (GroupSummary)}

\item{theme}{WebTheme object}

\item{interaction}{InteractionSpec object}
}
\description{
This is the central data structure that can be rendered as:
\itemize{
\item A forest plot (table + graphical interval column)
\item An interactive table (table only)
\item Other visualizations (upset plots, etc.)
}
}
