% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-core.R
\name{WebSpec}
\alias{WebSpec}
\title{WebSpec: Core specification for web-native table visualizations}
\usage{
WebSpec(
  data = (function (.data = list(), row.names = NULL) 
 {
     if (is.null(row.names)) {

            list2DF(.data)
     }
     else {
         out <- list2DF(.data,
    length(row.names))
         attr(out, "row.names") <- row.names
         out
     }

    })(),
  point_col = character(0),
  lower_col = character(0),
  upper_col = character(0),
  label_col = NA_character_,
  label_header = "Study",
  group_col = NA_character_,
  columns = list(),
  groups = list(),
  summaries = list(),
  overall_summary = GroupSummary(),
  scale = "linear",
  null_value = 0,
  axis_label = "Estimate",
  effects = list(),
  theme = NULL,
  interaction = NULL,
  labels = NULL,
  annotations = list()
)
}
\arguments{
\item{data}{Processed data as a data.frame}

\item{point_col}{Column name for point estimates}

\item{lower_col}{Column name for lower bounds}

\item{upper_col}{Column name for upper bounds}

\item{label_col}{Column name for row labels}

\item{group_col}{Column name for grouping (optional)}

\item{columns}{List of ColumnSpec objects defining table columns}

\item{groups}{List of GroupSpec objects}

\item{summaries}{List of GroupSummary objects}

\item{overall_summary}{Optional overall summary (GroupSummary)}

\item{scale}{Scale type: "linear" or "log"}

\item{null_value}{Reference value for null effect (0 for linear, 1 for log)}

\item{axis_label}{Label for the graphical axis}

\item{theme}{WebTheme object}

\item{interaction}{InteractionSpec object}
}
\description{
This is the central data structure that can be rendered as:
\itemize{
\item A forest plot (table + graphical interval column)
\item An interactive table (table only)
\item Other visualizations (upset plots, etc.)
}
}
